+++ packages/ui/src/ui/containers/ACL/ACL.tsx
@@ -9,19 +9,19 @@ import ColumnGroups from './ColumnGroups/ColumnGroups';
 import DeletePermissionModal from './DeletePermissionModal/DeletePermissionModal';
 import UserPermissions from './UserPermissions/UserPermissions';
 
-import LoadDataHandler from '../LoadDataHandler/LoadDataHandler';
-import ErrorBoundary from '../ErrorBoundary/ErrorBoundary';
-import DataTableYT from '../DataTableYT/DataTableYT';
-import MetaTable from '../MetaTable/MetaTable';
+import LoadDataHandler from '../../components/LoadDataHandler/LoadDataHandler';
+import ErrorBoundary from '../../components/ErrorBoundary/ErrorBoundary';
+import DataTableYT from '../../components/DataTableYT/DataTableYT';
+import MetaTable from '../../components/MetaTable/MetaTable';
 import {ClipboardButton, Loader, Popover} from '@gravity-ui/uikit';
-import Icon from '../Icon/Icon';
-import Link from '../Link/Link';
-import {Tooltip} from '../Tooltip/Tooltip';
-import {UserName} from '../UserLink/UserLink';
+import Icon from '../../components/Icon/Icon';
+import Link from '../../components/Link/Link';
+import {Tooltip} from '../../components/Tooltip/Tooltip';
+import {UserName} from '../../components/UserLink/UserLink';
 
 import withVisible, {WithVisibleProps} from '../../hocs/withVisible';
-import {renderText} from '../templates/utils';
-import Label from '../Label/Label';
+import {renderText} from '../../components/templates/utils';
+import Label from '../../components/Label/Label';
 import {isIdmAclAvailable} from '../../config';
 import ApproversFilters from './ApproversFilters/ApproversFilters';
 import ObjectPermissionsFilters from './ObjectPermissionsFilters/ObjectPermissionsFilters';
