+++ packages/ui/src/ui/containers/ACL/DeletePermissionModal/DeletePermissionModal.scss
