+++ packages/ui/src/ui/pages/accounts/tabs/acl/AccountsAclTab.js
@@ -7,7 +7,7 @@ import AccountsNoContent from '../../../../pages/accounts/AccountsNoContent';
 import {getActiveAccount} from '../../../../store/selectors/accounts/accounts';
 import {fetchAccounts} from '../../../../store/actions/accounts/accounts';
 import {loadUsers} from '../../../../store/actions/accounts/editor';
-import {AccountsAcl} from '../../../../components/ACL/ACL-connect-helpers';
+import {AccountsAcl} from '../../../../containers/ACL/ACL-connect-helpers';
 
 AccountsAclTab.propTypes = {
     activeAccount: PropTypes.string,
