+++ packages/ui/src/ui/pages/chyt/ChytPageClique/ChytPageCliqueAcl.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import {useSelector} from 'react-redux';
 
 import {getChytCurrentAlias} from '../../../store/selectors/chyt';
-import {AccessContentAcl} from '../../../components/ACL/ACL-connect-helpers';
+import {AccessContentAcl} from '../../../containers/ACL/ACL-connect-helpers';
 
 export function ChytPageCliqueAcl() {
     const alias = useSelector(getChytCurrentAlias);
