+++ packages/ui/src/ui/pages/tablet_cell_bundles/bundle/BundleAclTab.tsx
@@ -4,7 +4,7 @@ import ErrorBoundary from '../../../components/ErrorBoundary/ErrorBoundary';
 import {getTabletsActiveBundle} from '../../../store/selectors/tablet_cell_bundles';
 import {isPoolAclAllowed} from '../../../store/selectors/scheduling/scheduling';
 import {NoContent} from '../../../components/NoContent/NoContent';
-import {BundleAcl} from '../../../components/ACL/ACL-connect-helpers';
+import {BundleAcl} from '../../../containers/ACL/ACL-connect-helpers';
 
 export default function BundleAclTab() {
     const activeBundle = useSelector(getTabletsActiveBundle);
